# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gatewayservice API
  description: Gateway OpenAPI specification.
  version: 0.2.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://SOMEIP:8000
    description: Production server
paths:
  /adduser:
    post:
      summary: Add a new user to the database.
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: pass
      responses:
        '200':
          description: User added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    description: User ID
                  password:
                    type: string
                    description: Hashed password
                    example: $2b$10$ZKdNYLWFQxzt5Rei/YTc/OsZNi12YiWz30JeUFHNdAt7MyfmkTuvC
                  _id:
                    type: string
                    description: Identification
                    example: 65f756db3fa22d227a4b7c7d
                  createdAt:
                    type: string
                    description: Creation date.
                    example: '2024-03-17T20:47:23.935Z'
                  ___v:
                    type: integer
                    example: '0'                                        
        '400':
          description: Failed to add user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: getaddrinfo EAI_AGAIN mongodb
  /health:
    get:
      summary: Check the health status of the service.
      operationId: checkHealth
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Health status.
                    example: OK
  /login:
    post:
      summary: Log in to the system.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: pass
      responses:
        '200':
          description: Login successful. Returns user token, username, and creation date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: User token.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NWY3NTZkYjNmYTIyZDIyN2E0YjdjN2QiLCJpYXQiOjE3MTA3MDg3NDUsImV4cCI6MTcxMDcxMjM0NX0.VMG_5DOyQ4GYlJQRcu1I6ICG1IGzuo2Xuei093ONHxw
                  username:
                    type: string
                    description: Username.
                    example: student
                  createdAt:
                    type: string
                    description: Creation date.
                    example: '2024-03-17T20:47:23.935Z'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Shows the error info..
                    example: Invalid credentials
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /addgame:
    post:
      summary: Add a new game to the database.
      operationId: addGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gameName:
                  type: string
                  description: Name of the game.
                  example: Game1
      responses:
        '200':
          description: Game added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Message indicating success.
                    example: Game added successfully.
        '400':
          description: Failed to add game.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Game already exists.
  

  /getgamehistory/{username}:
    get:
      summary: Get game history for a specific user.
      operationId: getGameHistory
      parameters:
        - name: username
          in: path
          description: Username for which to retrieve game history.
          required: true
          schema:
            type: string
          example: student
      responses:
        '200':
          description: Game history retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    gameId:
                      type: string
                      description: ID of the game.
                      example: 12345
                    gameName:
                      type: string
                      description: Name of the game.
                      example: Game1
                    score:
                      type: integer
                      description: Score achieved in the game.
                      example: 100
                    datePlayed:
                      type: string
                      description: Date when the game was played.
                      example: '2024-03-20T15:00:00Z'
        '400':
          description: Failed to retrieve game history.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: User not found.

  /getScoreBoard:
    get:
      summary: Get the scoreboard.
      operationId: getScoreBoard
      responses:
        '200':
          description: Scoreboard retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    username:
                      type: string
                      description: Username of the player.
                      example: player1
                    score:
                      type: integer
                      description: Score achieved by the player.
                      example: 100

  /createquestion:
      get:
        summary: Get question from Wikidata
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    responseQuestionObject:
                      type: object
                      description: Question object
                    responseCorrectOption:
                      type: string
                      description: Correct option
                    responseAnswerOptions:
                      type: array
                      items:
                        type: string
                      description: Asnwer options
          '500':
            description: Error
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
        description: |
          Gets an random question from Wikidata and returns the question object, correct option and answer options.


  /getquestionshistory:
    get:
      summary: Get the history of generated questions.
      operationId: getQuestionsHistory
      responses:
        '200':
          description: Questions history retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    question:
                      type: string
                      description: The question.
                      example: What is 2 + 2?
                    answer:
                      type: integer
                      description: The correct answer to the question.
                      example: 4



  /getregisteredusers:
    get:
      summary: Get the list of registered users.
      operationId: getRegisteredUsers
      responses:
        '200':
          description: List of registered users retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    username:
                      type: string
                      description: Username of the registered user.
                      example: user1